convert multiple-event-handlers
